refactor(#106): 통합 테스트 중복 코드 개선 및 템플릿 작성
