+++ src/test/java/deepdivers/community/domain/hashtag/service/HashtagIntegrationTest.java
@@ -1,0 +1,7 @@
+package deepdivers.community.domain.hashtag.service;
+
+import deepdivers.community.domain.IntegrationTest;
+
+class HashtagIntegrationTest extends IntegrationTest {
+
+}
