+++ src/test/java/deepdivers/community/domain/image/application/ImageIntegrationTest.java
@@ -2,14 +2,14 @@ package deepdivers.community.domain.image.application;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import deepdivers.community.domain.ServiceTest;
+import deepdivers.community.domain.IntegrationTest;
 import deepdivers.community.domain.image.repository.entity.Image;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
-class ImageServiceTest extends ServiceTest {
+class ImageIntegrationTest extends IntegrationTest {
 
     @Autowired ImageService imageService;
