+++ src/test/java/deepdivers/community/domain/post/service/CommentIntegrationTest.java
@@ -3,7 +3,7 @@ package deepdivers.community.domain.post.service;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import deepdivers.community.domain.ServiceTest;
+import deepdivers.community.domain.IntegrationTest;
 import deepdivers.community.domain.common.NoContent;
 import deepdivers.community.domain.member.model.Member;
 import deepdivers.community.domain.post.dto.request.EditCommentRequest;
@@ -13,14 +13,13 @@ import deepdivers.community.domain.post.dto.request.WriteReplyRequest;
 import deepdivers.community.domain.post.dto.response.statustype.CommentStatusType;
 import deepdivers.community.domain.post.exception.CommentExceptionType;
 import deepdivers.community.domain.post.exception.PostExceptionType;
-import deepdivers.community.domain.post.model.comment.Comment;
 import deepdivers.community.domain.post.model.vo.CommentStatus;
 import deepdivers.community.global.exception.model.BadRequestException;
 import deepdivers.community.global.exception.model.NotFoundException;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
-class CommentServiceTest extends ServiceTest {
+class CommentIntegrationTest extends IntegrationTest {
 
     @Autowired CommentService commentService;
