+++ src/test/java/deepdivers/community/infra/aws/s3/S3TagManagerTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import deepdivers.community.domain.IntegrationTest;
 import deepdivers.community.global.config.LocalStackTestConfig;
 import deepdivers.community.global.exception.model.BadRequestException;
 import deepdivers.community.infra.aws.s3.properties.S3Properties;
@@ -23,22 +24,9 @@ import software.amazon.awssdk.services.s3.model.GetObjectTaggingResponse;
 import software.amazon.awssdk.services.s3.model.PutObjectRequest;
 import software.amazon.awssdk.services.s3.model.Tag;
 
-@SpringBootTest
-@Import(LocalStackTestConfig.class)
-@DirtiesContext
-@Transactional
-public class S3TagManagerTest {
+public class S3TagManagerTest extends IntegrationTest {
 
-    @Autowired
-    private S3TagManager s3TagManager;
-
-    @Autowired
-    private S3Client s3Client;
-
-    @Autowired
-    private S3Properties s3Properties;
-    @Autowired
-    private S3PresignManager s3PresignManager;
+    @Autowired private S3TagManager s3TagManager;
 
     @Test
     @DisplayName("S3에 저장된 객체에 태그를 설정한다.")
@@ -108,26 +96,4 @@ public class S3TagManagerTest {
             .hasFieldOrPropertyWithValue("exceptionType", NOT_FOUND_FILE);
     }
 
-    protected void createTestObject(String key) {
-        PutObjectRequest putObjectRequest = PutObjectRequest.builder()
-            .bucket(s3Properties.getBucket())
-            .key(key)
-            .build();
-
-        s3Client.putObject(putObjectRequest, RequestBody.fromString("test content"));
-    }
-
-    protected String generateAccessUrl(String key) {
-        return s3PresignManager.generateAccessUrl(key);
-    }
-
-    protected List<Tag> getTag(String key) {
-        GetObjectTaggingRequest getTaggingRequest = GetObjectTaggingRequest.builder()
-            .bucket(s3Properties.getBucket())
-            .key(key)
-            .build();
-
-        GetObjectTaggingResponse response = s3Client.getObjectTagging(getTaggingRequest);
-        return response.tagSet();
-    }
 }
